Update README.md
While we buy things on a daily basis, we tend to notice that a pattern evolves. We buy two packets of instant coffee every day and stuff like that.
With the insta - purchase app, a simple click of a button can add the required items to your e-cart on a shopping platform of your choice.