Update README.md
                                                                                Baseball handsign recognition app
In the game of basketball, coaches and pitchers often communicate with their on-field players with a variety of hand-signs. These hand-signs often can mean the difference between victory and defeat. So the opposing team tries to decrypt the hand signs made. 

a mobile application that takes into consideration the various signs and their outcomes can predict what may happen based on the pattern.

Input:
Hand- Symbols (or) signs

Output:
Pitcher's call.