Update README.md
                                                                                             Baseball Sign Predictor
In the game of basketball, coaches and pitchers often communicate with their on-field players with a variety of hand-signs. These hand-signs often can mean the difference between victory and defeat. So the opposing team tries to decrypt the hand signs made.

a mobile application that takes into consideration the various signs and their outcomes can predict what may happen based on the pattern.

Input:
Hand- Symbols (or) signs


Output:
Pitcher's call.
                                                                                              Music analysis
We often have a song at the top of our head and cannot really put on a finger on the name of the song. So, an application that can take into account the tone, rhythm and keywords and returns a list of probable songs can be revolutionary.

                                                                                             Insta - buy app
When we buy things from a shop on a daily basis, we tend to notice that certain things repeat. Like 2 packets of coffee every day. So an app where these can be added to an e-shopping site's cart at the push of a button can make lives much easier.